requirements.txt
# --------- pytorch --------- #
torch>=2.3.1
#torchvision>=0.11.0
pytorch-lightning==2.3.2
torchmetrics==1.4.0

# --------- huggingface --------- #
#transformers==4.42.3
datasets==2.20.0

# --------- hydra --------- #
hydra-core==1.3.2
hydra-colorlog==1.2.0
#hydra-optuna-sweeper==1.2.0  # for hyperparameter optimization

# --------- loggers --------- #
wandb
# neptune-client
# mlflow
# comet-ml

# --------- others --------- #
pyrootutils
#pre-commit
#rich
#pytest
#sh; platform_system != 'Windows'
omegaconf
tqdm
#tensorflow
#pynvml

# --------- data --------- #
numpy==1.26.4
soundfile==0.12.1
pandas

# --------- visualization --------- #
seaborn
scikit_learn
matplotlib
ipython
#grad-cam==1.3.1
#ttach==0.0.3
